# BuildManager: compute delta only for relevant projects #111