Previously the delta for the whole workspace was calculated even though
it was only checked for few projects. That caused a large overhead for
Workspaces with a lot projects.