@jukzi 2022-08-19
@jhonnen 
The computed delta of the whole workspace was never relay used - only for the relevant projects.
What's the (big) difference between DeltaDataTree.compareWith and DeltaDataTree.forwardDeltaWith? is it relevant?
I did not test this at all yet - just to get a idea and a CI test result