@jhonnen 2022-08-19
> What's the (big) difference between DeltaDataTree.compareWith and DeltaDataTree.forwardDeltaWith? is it relevant?

It probably has something to do with that part

```
	 * There is no requirement that a and b be related, although it is usually more
	 * efficient if they are. The node keys are used as the basis of correlation
	 * between trees.
	 *
	 * Note that if b is already represented as a delta over a,
	 * then c will have the same internal structure as b.
	 * Thus the very common case of previous forwardDeltaWith: current
	 * is actually very fast when current is a modification of previous.
```
... but I don't know what that means exactly, I'm not familiar with the delta computation.