@jukzi 2022-08-24
The current "solution" is sometimes slower then the previous code.