@jukzi 2022-08-24
There is something wrong in DeltaDataTree.basicCompare: the basicCompare loops over all tree.parent and calls searchNodeAt which also loops over the same trees. As the trees grows this is an O(n^2) for n trees:
![image](https://user-images.githubusercontent.com/51790620/186426796-6b84f269-a52b-4ef1-bbef-7ceeaf6ffb71.png)
The outer loop should continue at the tree where the inner loop found a result to avoid handling the same node over and over.