@jukzi 2022-10-12
Performance for my workspace with ~500 projects:
Each needsBuild() did cost about  ~00ms which is reduced to ~1ms.
The total full autobuild (upon changing the manifest) is reduced from 550sec to 420sec.
before:
![image](https://user-images.githubusercontent.com/51790620/195371681-27e6c774-20c3-4686-8318-d168fa480fb2.png)
after:
![image](https://user-images.githubusercontent.com/51790620/195372012-9d94e717-86b4-461c-9fae-d4492426105a.png)

It has a slight overhead if there is no delta, but that is irrelevant as both implementations are extremely quick (<1ms) when there is no delta.

In total it is a significant performance gain for autobuild with many projects in workspace after https://github.com/eclipse-platform/eclipse.platform/pull/186 was fixed