@iloveeclipse 2022-10-21
Please check https://github.com/eclipse-platform/eclipse.platform/issues/243 for a possible regression.